Aes128: Created the Aes128 library
The Aes128 library implements supports for encryption with the 128-bit
implementation of AES. It also implements support for an EncryptedDevice
in edevice. An EncryptedDevice is a wrapper for any BlockDevice, which
gives the user the option to encrypt data before writing it and decrypt
data after reading from a sector. The key in an EncryptedDevice can be
controlled by inputting a password.