VFAT: Create functions to allocate & free clusters
This adds 4 new functions:
* VFat::set_fat_entry(), replaces a FAT entry
* VFat::find_free_entry(), finds a free FAT entry
* VFat::alloc_cluster(), allocates a cluster with requested status
* VFat::free_cluster(), deallocates a cluster and marks it free

The goal here is to use these when modifying files. If we need to
extend the end of a file or create a new file, we call
vfat.alloc_cluster() with an EOC status. To extend the middle or
start of a file, we can call it with a Data status. If we want to
trim or delete a file we can use vfat.free_cluster().