VFAT: Implement VFat::write_cluster()
This is a really simple clone of VFat::read_cluster() with the buffers
swapped and a mut reference to the device instead of the buffer --
basically, should be the exact same to use as the existing
read_cluster() function. The goal here is to make the interface as
simple as possible and avoid as many dumb mistakes as possible. This
function will be used to allow us to write into files, where we will
enable seeking (just like in file reading) but we will only support
overwriting/appending to the end of a file. Again, the goal here is
simplicity, which is why we're not trying to create the most
general-purpose file writing mechanism here, that would allow e.g.
appending into the middle of the file. This means we'll need to have
entire files loaded in RAM, but it's not really a big deal because
FAT32 files should never be very large.

Make sure to extract the FAT32 images to test this :^)