VFat: Fix FAT entry count calculation in find_free_entry()
This is a big bug that slipped in the way I was doing find_free_entry()
before. Because I was calculating the wrong number of total FAT entries,
the code never reached all the FAT entries on the disk. In weird
circumstances it could probably go past the end of the FATs. The fix is
simple:

* Previously, I thought that the FATs were meant to be used in sequence
  rather than as duplicates. So no need to multiply by num_fats
* sectors_per_fat is not the # of sectors pointed to by a FAT, but
  literally the size of the FAT itself
* by multiplying sectors_per_fat by the number of entries per sector,
  we can calculate the total count of FAT entries