CachedPartition: Implement flushing cached writes to disk
Up until now, all writes in CachedPartition were happening to the
cache itself. This implements a flush() function for the cache that
iterates through the cached sectors and writes each dirty sector to
the disk. (To do this, I also had to implement the write_to_disk()
function). Finally, in order to enable flushing I implemented the
Drop trait for the CachedPartition, with a flush() occuring when
the cache is dropped. This means that if the system is safely
shutting down, the write will actually happen to the disk. This does
not protect us from unexpected shutdowns and in addition the kernel
will never actually drop the filesystem so this won't work as-is.

A good solution here might be adding a timer interrupt that tells
the OS to flush the cache. Alternatively, we could look into
scheduling flushes after every few writes (or even performing them
immediately). One final way of doing this would be with a
background kernel thread that handles the flush, but this would
take a little bit more work to get done.