VFAT: Implement write_chain_pos()
This function is more or less the reverse of read_chain_pos(). It
allows us to perform a write from an input buffer to the cluster
described by a Pos, following (and possibly extending) the file's
chain. Note that extending must occur here to grow files/directories
are more data is written to them.