VFAT: Implement seek_and_extend()
This function is a more advanced version of VFat::seek(), which is
meant to be used for seeking during a write. This function handles
the edge-case where we want to write to the end of a file or
directory, but the last cluster in the chain is already full. We
can't simply seek into the next (non-existent) cluster, so instead
we need to extend the chain and then seek into this newly added
cluster.