Dir/File: Keep track of actual entry for updating metadata
It's useful for files and directories to have pointers to their
directory entry for a number of reasons, such as modifying metadata
or deleting them. To make this happen, we need to modify the way
the directory iterator works, so that it keeps track of the current
entry's location on the disk. It's then able to pass this into the
File & Dir structs it creates so that you can access this later on.

Right now, this functionality isn't used for anything but it seemed
like a good time to add it before the Dir module grows much larger.