Dir: Implement file/directory creation
The Dir::create() function allows us to create new entries in a
directory based on the Metadata passed in. This handles both files
and directories and allows for normal names as well as LFNs. Quite
a bit of work has gone into these changes and it's likely going to
need a clean-up in the near future. For now, here's how it works:

Dir::create() calculates a start position for the first new entry.
This is done by finding the last used index in the dir and then
using VFat::seek_and_extend() to get the next position. After that
we also determine whether or not this entry is going to be LFN.

With that out of the way, we yield to either create_entry or
create_lfn_entry. The LFN entry handles allocating the extra LFN
entries, filling in the LFN-encoded string, and calculating the
checksum before yielding to the create_entry() function.

In create_entry(), all we have to do is encode the file's short
name, allocate some space for the first cluster (which we can
actually skip in the case of empty files, maybe we should do this
later...), and then write the entry to the disk. Finally, we return
back the proper directory entry so that the user can interact with
this new file.