Tests: Implement test_create_file()
This test creates an (LFN) file in the root directory and then checks
that it exists. Unfortunately, it's not very exhaustive and there's
lots of possible small cracks: metadata can be incorrect, LFN checksum
could be wrong, or it could have overwritten another file. For now,
the best bet is to actually mount the disk image and check it by hand.
For even better results, play with fsck.vfat + losetup and you'll get
to see all kinds of shortcomings of how this is implemented!

Additionally, this test doesn't clean itself up. After running the test
you should reset the disk image by running ./reset_tests.sh, which
resets the disk image so that you can be sure the test actualy worked.

Also note that to make this work, we had to make resources writeable.