File/Dir: Use Range instead of Pos to point to entry in parent
This is useful for LFN entries, as these may take the form of a long
string of entries in the parent directory. While we can just iterate
backwards in the simple case, it gets tough when going between
different clusters, because we don't have a reference to the previous
cluster.