FS: Implement file & directory deletion
This implements Dir::invalidate_entries() to (as the name implies)
invalidate all the entries in the specified range, effectively
deleting all traces of a file from the parent directory. Using this
API along with VFat::free_chain(), we implement File::delete() and
Dir::delete() to allow a user to delete a file/directory from its
handle -- note that Dir::delete() does not delete a file in its
directory, but the directory itself. Another thing to note about
Dir::delete() is that it doesn't handle deleting non-empty directories
and will return an Err if you attempt to do that.