Scheduler: Add with_running() to run a function over the running process
Since we can't clone processes and we can't return a borrowed value from
a GlobalScheduler function, the only way to let an external piece of code
modify processes is with lambdas. Here, we can take a function that tells
us what to do with the currently running process & modify it.