VM: Add PageTable::try_alloc()
PageTable::alloc() will panic if the entry already exists. This only
makes sense for the initial creation of a PageTable; after that, if
we try to allocate new pages there's a risk that the kernel might
panic. To get around this, we can now use PageTable::try_alloc(),
which unconditionally(*) returns a &mut [u8] to the physical location.

* If there's a serious problem we still panic