Process: Add load_existing() to load a program into an existing process
For the exec() syscall, we want to load a program into the currently
executing process. To do so, we need some way of performing
Process::load() and Process::do_load() on an existing process. This
function essentially performs both on the process (and uses the new
PageTable::try_alloc() function to avoid panics).