Syscalls: Add wait_pid() syscall for waiting on another process to die
The wait_pid() process is the final piece of the puzzle when it comes
to creating a userland shell: to run a binary from a user process, we
need to first fork() to create a child process, then load the program
with exec() and finally yield to the child process from the parent
using wait_pid().

The wait_pid() syscall is implemented by (per Will's great suggestion)
using an Arc<AtomicBool> stored in the process struct. Any process can
get this bool from the given process using SCHEDULER.get_dead_handle()
(note: this name sucks). The handle is ref-counted so we don't have to
worry about the process being dropped and the value inside is atomic
so we don't need to worry about any threading issues. Using this dead
handle, the current process is placed into the Waiting state until the
handle becomes true.

The user/echo program has been extended with this as well. Note that
the PID is now monotonically increasing after each command (since we
execute in the child). To return to the parent, simply type a '!' in
the console.