Syscalls: allow passing arguments through the exec() syscall
The exec() syscall now allows you to pass an &[&str] that contains
arguments to pass to the new program. This is a big behemoth of a
commit and I can't go over everything, but the majority of the work
happens in Process::init_args(). Other than that, we simply have to
marshal (marshall?) strings/slices between the user process & the
kernel's syscall interface (note: this will need some clean-up, I
basically went for the first thing that works rather than the most
resilient thing).

Essentially, args (truncated to ARG_MAX... 32 right now) are passed
on the stack after the new process is loaded (but before it's run).
Then, a pointer to the start of the args array is passed via the x1
register with the array's length passed in x0. For an example of how
to load these args into a program, check out user/echo/src/cr0.rs.

Side-note: I made fat32::util public because the SliceExt traits are
useful to have in the rest of the codebase.