Syscall: Add env_get() and env_set() syscalls
These syscalls allow us to interact with environment variables (and also
demo the general look we want to get out of the kernel API, such as how
to deal with strings & allocations). These will also be used to implement
pseudo chdir and getcwd syscalls (since IMO they kind of exist for no
reason).