Fs: Create FD module
The FD module implements file descriptors, which allow us to lock files,
map integers to file entries, and keep track of the owners of each file.
The architecture is pretty simple: a global FD table maps the FDs to
their locations/entries and keeps track of whether they're busy. All
lookups happen here, and using reference counting this table tracks the
number of unique FD owners (so we know when to actually close an opened
file for other processes to use).

A local FD table also exists and will be placed on each process to keep
track of the FDs that the specific process can access. Work is done here
to track changes in ownership in the global table, and the idea is that
child processes clone their parent's local FD table (thus incrementing
the ref count for the file) and each process can individually close or
use the file descriptor.