FAT32: Implement Entry::as_file_mut() and ::as_dir_mut()
These simply fill a big gap in the current implementation: we can
get an owned file/dir and a normal reference, but there was no way
to get a mutable reference to the file/dir (which is desirable when
the Entry isn't owned). Very useful for working with the FD tables,
which (as hash maps) will return a borrowed copy of the stored
Entry.