Syscalls: Add fs_open() and fs_close()
These are the equivalents of UNIX fopen() and fclose(). One current
thing to note is that fs_open() doesn't parse relative paths (because
we don't have the concept of a working directory in the kernel API).
Other than that, parse_path() tries to canonicalize the path name so
we don't get duplicate FDs for a busy file. Other than that, this is
just tying together loose ends in the API to get the FD API to the
user.