Syscalls: Add fs_create() syscall
This syscall is like UNIX's creat(). After playing around with a few
different designs, I decided upon this one: we don't require an FD to
a directory because that is a pain to deal with locked directories.
Internally, this uses an FD for the directory to lock it temporarily,
but thanks to the scheduler we can be sure that this won't cause any
issues (since nothing else can even request to open the directory at
the same time). Other than that, we choose not to use the Metadata
struct for creating files since most of the info isn't important when
creating a file (or necessarily something work exposing in an API).
Instead, we use the new EntryKind type which allows us to create a
file or directory, and we also specify the name of the new item.
Finally, this does some quick checks to make sure we don't create
duplicate dir entries under the same name.