Syscall: Add fs_delete() syscall
Like the fs_create(), we simply take as string path and internally try
to lock onto it to see if it's in use. If the file is not in use (and
if it's a directory, is empty) then we allow deletion. Right now
delete does some good error reporting so we can use this as a base for
improving error reporting in some of our other syscalls.