UserShell: Add support for a $PATH environment variable
Like the $PATH env var in UNIX, this points to the path of executables
(in this case /bin/) that we want the user to be able to call from
the shell. Since we don't yet have user-space allocation, all the
string concat code here is gross but we'll get around to fixing it
with format!() eventually (and support advanced features like allowing
multiple paths to be split with the : character).

To facilitate this change, we had to move all the binaries into /bin
and have the scheduler execute /bin/shell instead of /shell.