Page Faults: Kill process if it page faults in non-demand-paged memory
Right now, only the stack has demand-paging. If a program page faults
in any other region of memory, we don't allocate a new page. Therefore
this would cause an issue, since a page fault would loop forever for
that process. To get around that, we can simply kill the misbehaving
process so that it doesn't break the entire system. This is equivalent
to crashing from a stack overflow.