Syscalls: Implement request_page
The request_page syscall is analogous to sbrk() on UNIX. It allows
the user program to request a number of pages in virtual memory,
returning the previous program break (i.e. the first invalid
memory address). This works by keeping track of the last allocated
page by a process and incrementing it by the page size after each
new virtual memory allocation (excluding demand-paging for the stack,
since the stack isn't managed by the same system).