UserShell: Use user allocator (and fix obvious allocator bugs)
This sets us up with the GlobalAllocator trait properly, so we have
a single GlobalAllocator that's registered for the program. We were
required to bring in Mutex to get this working since unfortunately
the trait won't let us get away with a non-thread-safe construct like
Cell<T> (even though there's no multi-threading per allocator).