Page Faults: Avoid deadlock when killing misbehaving process
When killing misbehaving processes, we were calling SCHEDULER.kill,
which would try to acquire the lock for the scheduler. Unforunately,
this function was always called from within SCHEDULER.with_running
which is another critical section of the scheduler. As a result,
it would be impossible to ever acquire this lock and the program
would just hang instead of killing the user process. This simple
fix has the page fault return a bool so we can see whether we have
to kill in the trap handler.