Syscalls: Implement file_seek()
This syscall is analogous to UNIX's fseek() syscall. The implementation
is super simple, but we did need to write some convenience function to
deconstruct/reconstruct a shim::io::SeekFrom object. Unfortunately, we
can't export new impls on an existing type, so these are kind of ugly.
However, they hardly even need to be marked pub so it's not a big
concern of mine.