VFAT: Fix seek at end of cluster
When seeking from the end of a cluster, we previously only panicked
if we tried to seek more than 1 char into the EOC. However, this did
not take into account the fact that the loop already subtracted one
entire cluster of data, so in fact we would end up resetting the
position in the file to offset 0 in the current cluster. This meant
that the result from seek could not be trusted, and we relied on
seek_and_extend to repair our position.

Fixing this also exposed a bug in the implementation of File::seek,
since it could not seek past the end of a file to append when the
file ended at a cluster boundary. Much to the chagrin of Eric, this
bug has now been patched.